feat(SOLID 4): Liskov principle
Дети не должны менять параметры родителя.

Closes #614
