+++ src/solid/liskov/ShapeF.java
@@ -1,0 +1,13 @@
+package solid.liskov;
+
+/**
+ * Правильное получение {@link Squarer}
+ */
+public interface ShapeF {
+    /**
+     * Получение квадрата
+     *
+     * @return {@link Squarer}
+     */
+    int getSquarer();
+}
