+++ src/solid/liskov/package-info.java
@@ -1,6 +1,7 @@
 /**
  * <a href="https://github.com/Vachok/Java_Lessons/issues/614" target=_blank>Liskov Principle</a>
- *
+ * Основная концепция - нельзя менять поведение класса-родителя,
+ * работа должна осуществля
  * @since 01.06.2018 (16:41)
  */
 package solid.liskov;
